requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
pymongo==4.6.1
python-dotenv==1.0.0
requests==2.31.0
apscheduler==3.10.4
openai==0.28.0
dnspython==2.4.2
protobuf==4.24.4
tzdata==2023.3
watchdog==3.0.0